Moving the webhook handler class to the src folder
